build: add clang-tidy support
clang-tidy is a a powerful static analysis tool that comes with an
extensive set of checks to improve code quality and reliability.

This commit adds support for clang-tidy integration by adding a
.clang-tidy configuration file containing rules (and exceptions) this
project adheres to. These rules are enforced through the build process
using CMake which provides support for running clang-tidy using a
generated compile_commands.json file. Unfortunately, this absolutely
cripples build times so a dedicated configuration preset is added to
avoid running this all the time. Although not yet supported, this would
be a good candidate to run as part of the pre-commit process to both
verify that the project successfully compiles and passes all clang-tidy
validation checks. Additionally, getting this working on Windows
required implementing a workaround for a known defect with clang-tidy
as described here: https://gitlab.kitware.com/cmake/cmake/-/issues/20512

ReSharper C++ also has support for using a custom .clang-tidy file. See
https://www.jetbrains.com/help/resharper/Clang_Tidy_Integration.html for
more details.